+++ JANGAN/ExperimentQueue/Experiment3/UnbalancedDatasetAccuracy.ini
@@ -20,7 +20,7 @@ TextDownloadURLS = {
 	}
 
 # Deletes all the content in the folder specified above
-PurgePreviousData = True
+PurgePreviousData = False
 
 # Whether the output should contain numbers/letters
 # 1 or True for include | 0 or False for don't
@@ -46,7 +46,7 @@ LetterOutputFormat = "ZeroIndexed"
 ## To disable set both to 0
 ## Please don't set max below min
 MinimumLetterCount = 0
-MaximumLetterCount = 5000
+MaximumLetterCount = 100
 
 [CLASSIFIERDATAGENERATOR]
 BasePath = "../../Data"
@@ -61,7 +61,7 @@ TextDownloadURLS = {
 	}
 
 # Deletes all the content in the folder specified above
-PurgePreviousData = True
+PurgePreviousData = False
 
 # Whether the output should contain numbers/letters
 # 1 or True for include | 0 or False for don't
@@ -72,7 +72,7 @@ IncludeLetters = True
 # Path to distribution in given text files
 DistributionPath = "../../Data/Distribution.csv"
 # Whether to print distribution
-PrintDistribution = False
+PrintDistribution = True
 
 # Choses what format the output is 
 ## Letter - Output is the letter of the image
@@ -87,7 +87,7 @@ LetterOutputFormat = "ZeroIndexed"
 ## To disable set both to 0
 ## Please don't set max below min
 MinimumLetterCount = 0
-MaximumLetterCount = 5000
+MaximumLetterCount = 100
 
 [CGANTRAINING]
 # Directories for the datasets
@@ -97,7 +97,7 @@ TestDatasetDir = ""
 ImageSize = 28
 BatchSize = 32
 LatentDimension = 128
-EpochCount = 50
+EpochCount = 2
 # How often to refresh the UI under training
 #    Note: low value will slow the training down
 RefreshUIEachXIteration = 30
@@ -130,7 +130,7 @@ FormatImages = True
 OutputDir = "../../Data/{EXPERIMENTNAME}/{TIMESTAMP}/FakeOutput/Final/"
 # How many images the trained generator should make
 # It generates this amount of images for each type
-NumberOfFakeImagesToOutput = 1000
+NumberOfFakeImagesToOutput = 10
 
 [CLASSIFIERTRAINING]
 # Directories for the datasets
