+++ JANGAN/ExperimentQueueConfig.ini
@@ -3,7 +3,7 @@
 ThrowIfConfigFileBad = True
 #ExperimentList = [ "BASERUN" ]
 # These are for Experiment 3:
-ExperimentList = [ "UNBALANCEDDATASETACCURACY", "UNBALANCEDDATASETQUALITY" ]
+ExperimentList = [ "UNBALANCEDDATASETACCURACY" ]
 
 # Name of experiment
 #    ModuleName: Name of the override module to import from the ExperimentQueue folder
@@ -33,9 +33,9 @@ ClassifyImages = True
 ModuleName = "ExperimentQueue.Experiment3.UnbalancedDatasetAccuracy"
 AmountOfTimesToRun = 1
 ConfigFile = "ExperimentQueue/Experiment3/UnbalancedDatasetAccuracy.ini"
-MakeCGANDataset = True
-MakeClassifierDataset = True
-TrainCGAN = True
+MakeCGANDataset = False
+MakeClassifierDataset = False
+TrainCGAN = False
 ProduceCGANLetters = True
 TrainClassifier = True
 ClassifyImages = True
